Takes the differences between x and the reference values

Takes the differences between x and the reference values

drift_reference.differentiate(x, timestamps, reference, scalefactor.sec)

Value

A data.frame (x, timestamps, reference.id) with the differences. Note that in case of more than 1 reference is supplied, there might be duplicate timestamps based on different reference barometers.