gwloggeR: Validate and annotate raw logger air and hydrostatic pressure time series

Groundwater divers measure hydrostatic pressure, air pressure and temperature. These raw time series contain issues such as outliers, drift,... This package provides functions to facilitate the validation workflow of groundwater related time series. Each of the functions takes a time series and annotates problematic observations with a controlled vocabulary of flags.

Details

TODO: package details.

References

GitHub

See also

Author

Maintainer: Davor Josipovic davor.josipovic@inbo.be (ORCID)