Detect outliers

This function marks outliers in the input vector.

detect_outliers(x, apriori, ..., plot = FALSE, verbose = FALSE,
  title = NULL, timestamps = NULL)

# S4 method for numeric,missing
detect_outliers(x, plot, verbose, title,
  timestamps)

# S4 method for numeric,apriori
detect_outliers(x, apriori, plot, verbose,
  title, timestamps)

Arguments

x
numeric vector of values

apriori
apriori class

...
optional parameters, depending on signature:

plot
prints comprehensive plots

verbose
prints comprehensive information

title
adds title to the plot

timestamps
timestamp vector. For airpressure, timestamps are of no importance, except aestehtical for the scatter plot if plot = TRUE. In case there are duplicates and NA values, only warnings will be raised which might suggest that something is wrong with x. For hydrostatic pressure, timestamps are important. Therefore an error is raised if timestamps are not supplied, or if any of the timestamps are NA. Note that in any case, timestamps do not have to be ordered.

Value

Logical vector with same length as x, specifying TRUE for an outlier.

Methods (by class)

x = numeric,apriori = missing: Only considers x, without any a-priori information. A normal distribution is assumed with mean and variance estimated using median and MAD as described in Leys, 2013.

x = numeric,apriori = apriori: Takes a-priori information about x into consideration.

References

Chen, C. e.a., Joint Estimation of Model Parameters and Outlier Effects in Time Series, 1993. Leys, C. e.a., Detecting outliers, 2013.

Examples

# In case of a vector:
x <- c(1:10, 100)
detect_outliers(x)

#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE

# In case of a dataframe, select the column:
detect_outliers(cars$dist)

#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [49] FALSE FALSE

# Or use the tidyverse approach:
library(magrittr)
cars %>% dplyr::mutate("outlier" = detect_outliers(dist))

#>    speed dist outlier
#> 1      4    2   FALSE
#> 2      4   10   FALSE
#> 3      7    4   FALSE
#> 4      7   22   FALSE
#> 5      8   16   FALSE
#> 6      9   10   FALSE
#> 7     10   18   FALSE
#> 8     10   26   FALSE
#> 9     10   34   FALSE
#> 10    11   17   FALSE
#> 11    11   28   FALSE
#> 12    12   14   FALSE
#> 13    12   20   FALSE
#> 14    12   24   FALSE
#> 15    12   28   FALSE
#> 16    13   26   FALSE
#> 17    13   34   FALSE
#> 18    13   34   FALSE
#> 19    13   46   FALSE
#> 20    14   26   FALSE
#> 21    14   36   FALSE
#> 22    14   60   FALSE
#> 23    14   80   FALSE
#> 24    15   20   FALSE
#> 25    15   26   FALSE
#> 26    15   54   FALSE
#> 27    16   32   FALSE
#> 28    16   40   FALSE
#> 29    17   32   FALSE
#> 30    17   40   FALSE
#> 31    17   50   FALSE
#> 32    18   42   FALSE
#> 33    18   56   FALSE
#> 34    18   76   FALSE
#> 35    18   84   FALSE
#> 36    19   36   FALSE
#> 37    19   46   FALSE
#> 38    19   68   FALSE
#> 39    20   32   FALSE
#> 40    20   48   FALSE
#> 41    20   52   FALSE
#> 42    20   56   FALSE
#> 43    20   64   FALSE
#> 44    22   66   FALSE
#> 45    23   54   FALSE
#> 46    24   70   FALSE
#> 47    24   92   FALSE
#> 48    24   93   FALSE
#> 49    24  120   FALSE
#> 50    25   85   FALSE